+++ examples/cobra/main.go
@@ -30,5 +30,5 @@ func init() {
 }
 
 func main() {
-	rootCmd.Execute()
+	_ = rootCmd.Execute()
 }