+++ examples/msg.go
@@ -4,4 +4,4 @@ package examples
 type Msg string
 
 // Entrypoint is a dummy implementation of the method "Entrypoint of the dummy engine"
-func (s Msg) Entrypoint() string { return "" }
+func (s Msg) Entrypoint() string { return string(s) }