+++ handlers/opentracing/tracing_test.go
@@ -21,7 +21,7 @@ var (
 
 func makeTracerContext(i int) *engine.TxContext {
 	// Initialize context
-	txctx := engine.NewTxContext().Prepare([]engine.HandlerFunc{}, log.NewEntry(log.StandardLogger()), nil)
+	txctx := engine.NewTxContext().Prepare(log.NewEntry(log.StandardLogger()), nil)
 
 	txctx.Reset()
 	switch i % Mod {